refactor: Move enemy related things to the bullet module
Visibility changes to `pub`:
- some fields in `EnemyStateShared`
- fields in `Enemy`
